Movie Recommendation System using Java
A system in Java that provides recommendations to users based on the movies they watch, their friends, watch history, ratings by film critics, etc. Done under the guidance of faculty from the Department of Computer Science, Duke University as part of the Java Programming: An Introduction to Software Coursera specialization.